+++ libs/Meact/jsx-runtime.ts
@@ -1,22 +1,4 @@
-type DomNode = HTMLElement;
-
-type JsxsConfig = {
-  [key: string]: Element[];
-};
-
-type JsxConfig = {
-  [key: string]: string | number | Element;
-};
-
-type ElementKey = string | number;
-
-type Element = {
-  type: string;
-  key: string | null;
-  props: JsxsConfig | JsxConfig;
-};
-
-export const Fragment = "fragment";
+import { ElementKey, JsxConfig, JsxsConfig } from "../../type/meact";
 
 const makeElement = (
   type: string | Function,
@@ -65,32 +47,4 @@ export const jsx = (
   return makeElement(type, config, elementKey);
 };
 
-export const render = (meactNode: Element, domNode: DomNode) => {
-  const dom = document.createElement(
-    meactNode.type === Fragment ? "div" : meactNode.type
-  );
 
-  Object.keys(meactNode.props).forEach((key) => {
-    if (key !== "children" && meactNode.props[key]) {
-      dom.setAttribute(key, meactNode.props[key].toString());
-    }
-  });
-
-  switch (typeof meactNode.props.children) {
-    case "string":
-    case "number":
-      dom.textContent = meactNode.props.children.toString();
-      break;
-
-    case "object":
-      if (Array.isArray(meactNode.props.children)) {
-        meactNode.props.children.forEach((child: Element) =>
-          render(child, dom)
-        );
-      } else {
-        render(meactNode.props.children as Element, dom);
-      }
-      break;
-  }
-  domNode.appendChild(dom);
-};
